Refactor navigation in our app
