+++ lib/features/home/presentaion/screen/home_screen.dart
@@ -1,8 +1,8 @@
 import 'package:bookly_app/core/utils/styles.dart';
-import 'package:bookly_app/view/screens/widgets/home_screen/best_seller/best_seller_listview_widget.dart';
+import 'package:bookly_app/features/home/presentaion/screen/widgets/best_seller/best_seller_listview_widget.dart';
 import 'package:flutter/material.dart';
-import 'widgets/home_screen/Row_appbare_Widget.dart';
-import 'widgets/home_screen/list_view_widget.dart';
+import 'widgets/Row_appbare_Widget.dart';
+import 'widgets/list_view_widget.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
