+++ lib/features/splash_screen/splash_screen.dart
@@ -1,4 +1,4 @@
-import 'package:bookly_app/view/screens/widgets/splash_screen/splash_body_widget.dart';
+import 'package:bookly_app/features/splash_screen/screen/widgets/splash_body_widget.dart';
 import 'package:flutter/material.dart';
 
 class SplashScreen extends StatelessWidget {
