+++ pubspec.yaml
@@ -30,7 +30,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  get: ^4.6.5
+  go_router: ^6.0.9  
   google_fonts: ^4.0.3
   font_awesome_flutter: ^10.4.0
